[#10] Add flags to configure secure connection client certificate
- implementation only
