+++ internal/feature_test.go
@@ -18,15 +18,11 @@ import (
 	"testing"
 
 	"github.com/eclipse-kanto/software-update/hawkbit"
-	"github.com/eclipse-kanto/software-update/internal/storage"
 )
 
 const (
-	testDirFeature        = "_tmp-feature"
-	testDefaultHost       = ":12345"
-	testDefaultHostSecure = ":12346"
-	testCert              = "storage/testdata/valid_cert.pem"
-	testKey               = "storage/testdata/valid_key.pem"
+	testDirFeature  = "_tmp-feature"
+	testDefaultHost = ":12345"
 )
 
 // TestScriptBasedConstructor tests NewScriptBasedSU with wrong broker URL.
@@ -85,7 +81,7 @@ func TestScriptBasedInitLoadDependencies(t *testing.T) {
 
 	// 1. Try to init a new ScriptBasedSoftwareUpdatable with error for loading install dependencies
 	_, _, err := mockScriptBasedSoftwareUpdatable(t, &testConfig{
-		clientConnected: true, storageLocation: dir, featureID: getDefaultFlagValue(flagFeatureID)})
+		clientConnected: true, storageLocation: dir, featureID: defaultFeatureID})
 	if err == nil {
 		t.Fatalf("expected to fail when mandatory field is missing in insalled dept file")
 	}
@@ -100,14 +96,14 @@ func TestScriptBasedInit(t *testing.T) {
 
 	// 1. Try to init a new ScriptBasedSoftwareUpdatable with error for not connected client
 	_, _, err := mockScriptBasedSoftwareUpdatable(t, &testConfig{
-		clientConnected: false, storageLocation: dir, featureID: getDefaultFlagValue(flagFeatureID)})
+		clientConnected: false, storageLocation: dir, featureID: defaultFeatureID})
 	if err == nil {
 		t.Fatal("Ditto Client shall not be connected!")
 	}
 
 }
 
-// TestScriptBasedCore tests ScriptBasedSoftwareUpdatable core functionality: init, install and download.
+// TestSBSUInit tests ScriptBasedSoftwareUpdatable core functionality: init, install and download.
 func TestScriptBasedCore(t *testing.T) {
 	// Prepare
 	dir := assertPath(t, testDirFeature, false)
@@ -115,33 +111,23 @@ func TestScriptBasedCore(t *testing.T) {
 	defer os.RemoveAll(dir)
 
 	// Prepare/Close simple HTTP server used to host testing artifacts
-	w := storage.Host(testDefaultHost, "", 0, true, false, "", "", t).AddInstallScript()
-	defer w.Close()
-
-	wSecure := storage.Host(testDefaultHostSecure, "", 0, true, true, testCert, testKey, t).AddInstallScript()
-	defer wSecure.Close()
+	w := host(testDefaultHost, t).addInstallScript()
+	defer w.close()
 
 	// 1. Try to init a new ScriptBasedSoftwareUpdatable.
 	feature, mc, err := mockScriptBasedSoftwareUpdatable(t, &testConfig{
-		clientConnected: true, featureID: getDefaultFlagValue(flagFeatureID), storageLocation: dir})
+		clientConnected: true, featureID: defaultFeatureID, storageLocation: dir})
 	if err != nil {
 		t.Fatalf("failed to initialize ScriptBasedSoftwareUpdatable: %v", err)
 	}
 	defer feature.Disconnect()
 
-	testDownloadInstall(feature, mc, w.GenerateSoftwareArtifacts(false, "install"), t)
-
-	feature.serverCert = testCert
-	testDownloadInstall(feature, mc, wSecure.GenerateSoftwareArtifacts(true, "install"), t)
-}
-
-func testDownloadInstall(feature *ScriptBasedSoftwareUpdatable, mc *mockedClient, artifacts []*hawkbit.SoftwareArtifactAction, t *testing.T) {
 	// Preapare simple software update action.
 	sua := &hawkbit.SoftwareUpdateAction{
 		CorrelationID: "test-correlation-id",
 		SoftwareModules: []*hawkbit.SoftwareModuleAction{{
 			SoftwareModule: &hawkbit.SoftwareModuleID{Name: "test", Version: "1.0.0"},
-			Artifacts:      artifacts,
+			Artifacts:      w.getSoftwareArtifacts("install"),
 			Metadata:       map[string]string{"artifact-type": "plane"},
 		}},
 	}
