--- internal/storage/testdata/expired_cert.pem
@@ -1,23 +1,0 @@
------BEGIN CERTIFICATE-----
-MIID3TCCAsWgAwIBAgIUNHqJ37Ms9XbUj4roRGbXXyn/nmowDQYJKoZIhvcNAQEL
-BQAwfjELMAkGA1UEBhMCVVMxEzARBgNVBAgMCkNhbGlmb3JuaWExFjAUBgNVBAcM
-DU1vdW50YWluIFZpZXcxGjAYBgNVBAoMEVlvdXIgT3JnYW5pemF0aW9uMRIwEAYD
-VQQLDAlZb3VyIFVuaXQxEjAQBgNVBAMMCWxvY2FsaG9zdDAeFw0yMjAxMjgxNTAw
-MDBaFw0yMjAxMjkxNTAwMDBaMH4xCzAJBgNVBAYTAlVTMRMwEQYDVQQIDApDYWxp
-Zm9ybmlhMRYwFAYDVQQHDA1Nb3VudGFpbiBWaWV3MRowGAYDVQQKDBFZb3VyIE9y
-Z2FuaXphdGlvbjESMBAGA1UECwwJWW91ciBVbml0MRIwEAYDVQQDDAlsb2NhbGhv
-c3QwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDJQQSb/svm4Bg0AZ93
-JzyXkZ80Yn2fLsdP9QXhMravBoTNm36xsq38HEVUfb7b1UJdT8eS7PNdbUAHuFoE
-qFwlSA7nBq1n5ihrHwzyP6rm76N9jVopwCUdYz56vsGUQeiv01MIkWNvlRtYyrr5
-qhZ1kPPnNnNG2TTiSFcF0+OpXUFfu5HIImxgqo5nWboYKCiVR9J6fEkp44oQPuCe
-JxuXwybL/w56J8ZHqfSh4l7v2tulVeBxLXoAgDkKuAWZ6lVHXx+1MjfEaaJ9YB8h
-FGNiXFRdcnZFmHBXFXBCiAj8BL57cZd2GKZxsotb9unxSnt6GV+op9wNI0onDo2X
-pQUpAgMBAAGjUzBRMB0GA1UdDgQWBBTqgTcithahyOF2QEhrM0xu3zywGTAfBgNV
-HSMEGDAWgBTqgTcithahyOF2QEhrM0xu3zywGTAPBgNVHRMBAf8EBTADAQH/MA0G
-CSqGSIb3DQEBCwUAA4IBAQCJA5aOzv9evIXoTI3infqICSAU4gWf9Pou5wgjNYG6
-ZcaztyzvPbxY0HtjeLhzA9Yx3lQR0KahWBaQ9uLMDZS4RyWZUE/4Wddw4tAWscaH
-iS1qpxcfD8jns3Geyr02y5vhiJtER9r9vNg2/hZThuW730x4zinXObCBYklMPJJr
-iJ8wzfovFgAev5OOtyB8v0nPPmAffLI19tcGQyOj/rVO97mmKPoAYpykvW9dGlQU
-0by4MmFHIHOYqeDtufO3/kkQq1GS86EzKn0YGVjLa41gqMw9kw043QKJqdrbn8tS
-OJVgOfXaAfTjvwrNPemmK3YWk5uqesmUenFzRh/KCrjo
------END CERTIFICATE-----
