--- internal/storage/testdata/expired_key.pem
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvwIBADANBgkqhkiG9w0BAQEFAASCBKkwggSlAgEAAoIBAQDJQQSb/svm4Bg0
-AZ93JzyXkZ80Yn2fLsdP9QXhMravBoTNm36xsq38HEVUfb7b1UJdT8eS7PNdbUAH
-uFoEqFwlSA7nBq1n5ihrHwzyP6rm76N9jVopwCUdYz56vsGUQeiv01MIkWNvlRtY
-yrr5qhZ1kPPnNnNG2TTiSFcF0+OpXUFfu5HIImxgqo5nWboYKCiVR9J6fEkp44oQ
-PuCeJxuXwybL/w56J8ZHqfSh4l7v2tulVeBxLXoAgDkKuAWZ6lVHXx+1MjfEaaJ9
-YB8hFGNiXFRdcnZFmHBXFXBCiAj8BL57cZd2GKZxsotb9unxSnt6GV+op9wNI0on
-Do2XpQUpAgMBAAECggEBAJBjWFY/ALU8s8CkJW+MM/meXXVs8CnDmvWvmi9Ofiim
-aaNF2yjMQ+0rr0YGDWeU4k8mDM5Jt/SgjB690b+jtrB8uUxgywpO41KvbU9kBE45
-QbezMSvhEf+6kznnH+UpQDLdT6ZHAOsuM/eJQ4oZ0JE6lPKeayg6b6mlMuIYei6o
-VLl4Iupdn34sIJidwbMzlWdgndh1dlR0DWob/J/e1DWRQxJ6Rv1tppV6bPPmU8ee
-8gMv7VOruG2YOkhuKjtIZMeVQHI71cN/zrB66Ma8xQmW4lbVS0tuURpK4mfUNhuK
-Vnnp1EgsbHkKy4xeyCSyIS979u5Qbq8P08pJKVhSju0CgYEA/SthtyYHirqoukX8
-l9pdKtGifuM7acdWrQY5R+fKL+64KOWGBGyQaGYcXolZTB4F9rOcdHR/0l6FIU9d
-QJBnPHvU/MFRs02sukmnHW5Pe7LUU3aqgfs6NOrvQKen2Mx3YeCLIh4WRYt9FzhO
-mqBCMGO8IWPipfRJriNShfWMdyMCgYEAy4ELY8OJk5caFIZsuE5ODPGrQSZIFRiy
-kAec+A28Td+ZzOUQ8lwZWelHZylbleeArdrfpUcCA1Wk1Dnc7405gdftCD/o0KX9
-EhooMIeA2qY6E0MibproJuos8cF8S94ToCvZsJCUktcQ3j4uidO+u1t1/2MpAWr1
-FtP2uYWvvUMCgYEAvaCA0G3ruD2Xo+5IBHXK0weybthE6DYm9fJzdShoiP6Fchx5
-BVdL0uao2SaHfDwAu4hM5+1a0Hd7wDdx4apHqUYXgjpKt8K2XdJNGf5GCqEn6Fw5
-2qISNxDBo+vwDLMZuHY3wJZImE6io4HJXogrC5Oolln6qdMQXR6X8T66t4ECgYEA
-ocIQSJt79rnrvBbhwlpTxgwuwRlpwIMiJEQ47B8W55ttiCUoJE06xXIdFI4WQ4CB
-ky/gB2L64aWo4ByiVjOeW2xD+AO8zmSyVTo/adu7e5EXpon9i4dMMup/ESdn4nJC
-jVVtsuApiNK0kMNCgMGz4P8hOwAZ7wuShzOy8OKnvzsCgYBGb3NLiP1YsgpJWqzt
-nIYzgs9wP8jWc02jeqWQToz1hDCi+ZY0SzZ/yxhH+51ItPW5/e0kY90dTRZL1wob
-zc6TsRwf1V5L9WTX8FhNLeA+ECfg5RXzVvnvPEdeQXKjzoCnf/DWCSrjRnhf9OyE
-ThfFFGdrtZdjQzVadRB/ViiB+w==
------END PRIVATE KEY-----
