--- internal/storage/testdata/ext_cert.pem
@@ -1,23 +1,0 @@
------BEGIN CERTIFICATE-----
-MIID3TCCAsWgAwIBAgIUY3OuDRo1Xrrt255EhxPVybPKfH8wDQYJKoZIhvcNAQEL
-BQAwfjELMAkGA1UEBhMCVVMxEzARBgNVBAgMCkNhbGlmb3JuaWExFjAUBgNVBAcM
-DU1vdW50YWluIFZpZXcxGjAYBgNVBAoMEVlvdXIgT3JnYW5pemF0aW9uMRIwEAYD
-VQQLDAlZb3VyIFVuaXQxEjAQBgNVBAMMCWxvY2FsaG9zdDAeFw0yMjAyMDMwNzMx
-NDdaFw0zNTEwMTMwNzMxNDdaMH4xCzAJBgNVBAYTAlVTMRMwEQYDVQQIDApDYWxp
-Zm9ybmlhMRYwFAYDVQQHDA1Nb3VudGFpbiBWaWV3MRowGAYDVQQKDBFZb3VyIE9y
-Z2FuaXphdGlvbjESMBAGA1UECwwJWW91ciBVbml0MRIwEAYDVQQDDAlsb2NhbGhv
-c3QwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDwCCv4PtcPJ2a6bQl4
-KHrbi4bFhx8SYG56C5ZvPvtB6caD4Q+eH1P8T6Y1qsBWN4/EQ377GsnSRlqgigpT
-HhRXw5Wsm0NIFOBTLepV2u6gH++Mm2VLUF7RETZFjIhn/ouINTa2gy+bej3/AD8O
-yTa3f1vyxkPk6SRu6DUs1X6u/vqaoeMWhlpn3/XDh7OSgVlXohBe5jWLr0Y9jiHk
-EmMVSTbqyVy5micFN/VlSiaWDOH9CDCtu6E8Z7hUpmGFgKrKUWBadvGC9PnYewdo
-pWbILPlTBcKHno3kn3I8mU1iW4XXRQbcRo4MfuGwFCUvwSZb2I6SeV4kmNHa0IM/
-YWlFAgMBAAGjUzBRMB0GA1UdDgQWBBStUNPdRYiihwJv08dbHl/U3paWJjAfBgNV
-HSMEGDAWgBStUNPdRYiihwJv08dbHl/U3paWJjAPBgNVHRMBAf8EBTADAQH/MA0G
-CSqGSIb3DQEBCwUAA4IBAQC1TIJTjlvmVAoRAhNdkCM8aNJrfaeBKyIwy9J7hnzu
-wU5EpJsRkiaaNsXLwop/RwT7Hp/D5RVzTpGWpUUW78LC5d2bRKbTvmjH0hlkEZvt
-K6WYKaXITh/xgkVnQuJ6KOGX6cTcAf6C2GdLQAMgYoJdxPAGXNk06Uoif4snYP93
-gBREV7fcKU0n5ZDc5lBFDeKwpa9hXS+auMnj7K2Vb/UlQjC0RHijhsPE7CxPWvJ3
-sE7Qfi8ulTivP+HoqS1DmzDwUnIszABIUmWoLqDPc8Jmmx0/hMYT088yMfhccTxd
-QBJpCvSdIEjQy2pLWWWsjn6lj5+2i+RE1V+0qyFjX5dw
------END CERTIFICATE-----
