--- internal/storage/testdata/ext_key.pem
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDwCCv4PtcPJ2a6
-bQl4KHrbi4bFhx8SYG56C5ZvPvtB6caD4Q+eH1P8T6Y1qsBWN4/EQ377GsnSRlqg
-igpTHhRXw5Wsm0NIFOBTLepV2u6gH++Mm2VLUF7RETZFjIhn/ouINTa2gy+bej3/
-AD8OyTa3f1vyxkPk6SRu6DUs1X6u/vqaoeMWhlpn3/XDh7OSgVlXohBe5jWLr0Y9
-jiHkEmMVSTbqyVy5micFN/VlSiaWDOH9CDCtu6E8Z7hUpmGFgKrKUWBadvGC9PnY
-ewdopWbILPlTBcKHno3kn3I8mU1iW4XXRQbcRo4MfuGwFCUvwSZb2I6SeV4kmNHa
-0IM/YWlFAgMBAAECggEAd22ZcwrRNlkEY/2cmIyLlHD3vZBdhWc1kMPzNabAnSsT
-+sogWyx7PS7kJA+xYzSB88UhiQ7rHBSlQvtr568vr/tAnYEjGNpKQxi1X8JuuH0r
-U2p+G9sBUkSeHDywIg4knsDbWDQtm85Wymn09gZnkyVrOoArh9n4prYivwv63FZK
-40t+O5FOc1yc7QQ+qXQXkvvBaCExiZaPm2+nNj6aBPTfvZY7whdLY33wjbk7n5GO
-7xroZoIuzWnoglmyammVF7E+Btd412XJ2AS9jY47Pf8jGpJNG3vdEqRZrGu4etiF
-AO57WSMNBcBgWdBmihiFLgsZ1if59gcdre8gfFvYgQKBgQD8Sf56FytfBpZ5KEuf
-KlsUeUW+uJErVuTXWZ8agfZvuhITMWVzPyQOEsAeBJXP047kmL/7+AGJJW/NBulH
-7OQpUmh8iQHHBzsB/QsmKZ61cafzKpFt3X9HIWvgf7+EZ1ulqxaJSESXlL2Y1GNS
-gRkFQNGWU2IYG0VE58+fKP46mwKBgQDzkAXhB0XM3c8C7+OOZT0fYVllNwSycYYn
-JkNrk44hjdikNRKRG1kZveN1z2soly17wOhN0kifoTNJqy18jBTY4qGfS6LkRjnV
-NIIM0OV2NCVMTLbYc0jJiajLo6+KX5ujLQiVote1VLq0bcXsWb/gPNCkk/3F7S42
-3UwlUPO5nwKBgB6kfhwmJEZubmhY1JaIdneOI0J++GLjMvhVElYN4vi/COL75O6S
-hLaQ4HTE6c1MTNbHFspr3QPjYCju5faudyftsnMt0OFzp18edPQ1qekv3M5UJUXK
-2WKyVH/4BDL4utDsbthNxscCceTWJxomQ/u3W71+czOqXEd6QGV8fSRxAoGAfFYZ
-qja8O0lAVI+B6z9jZJGdwT3E+BNcGMA/yrsY46g14B33zHvZkHqPk6S/X4lNfs+L
-x66BxHvWYtUv81TPj6iiOV+zWDVBcMXNXh2GBkG4iKrg7L5tefT8Zni5NDp6iWKD
-71+iHr+SUGl5kJ6AhUiGen3SHyhlXzy5kpMw2KECgYEAqWshX4nYzjYyQt7QrJWz
-K9mBQOWtbmOOLZZwmoIypBbmYf+sRqIfhbVdjTd/tKXMKceqz2q1AuGHzpMOLeIA
-tD/8veIazVHBhNarCeZh5lf1ViTyft/TOnl6JT/6YzZdlY5jPWLtxUen9LAlMStx
-kIDHtJX9qtRHCcBQQz7j0o0=
------END PRIVATE KEY-----
