--- internal/storage/testdata/valid_cert.pem
@@ -1,30 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFJTCCAw2gAwIBAgIUAVys7/gWAuMqTRbmGf1PzyIVh/owDQYJKoZIhvcNAQEL
-BQAwFDESMBAGA1UEAwwJbG9jYWxob3N0MB4XDTIyMDgxNjEyMDIxOFoXDTMyMDgx
-MzEyMDIxOFowFDESMBAGA1UEAwwJbG9jYWxob3N0MIICIjANBgkqhkiG9w0BAQEF
-AAOCAg8AMIICCgKCAgEAxaQ6nBtwTqmPZDb+FQwq8NrIYK1MPw0E9CmFIYt5liLa
-BoXzxd1H65o+fmhho4RzWbdxmBTqcXcJ6u9A8UTAl2P/1COp6YSssVNzSaqIOpGg
-SeHCEe631Qfn6wjDQPh2SOJa0CQ4B848ox9cyIfsGJzNLKzxBcNm94KQ/nys/wad
-+6BLEivG53JhT6Sf4pePMZWErD0zzlJrpHC+/gp1y3ubfa8m+YWwz89lgNTKt6KO
-XxxCu3QDCUJEZOZzUEmsPVPptJV6NmE3MwoB4bA3r2fNdyxvzINcXSPfy/i1b/Zh
-w9uzWKRaHJTaVp/pytH5Fzw6RYQ2rYV+gPAFWbh/i6G0TKxqbAprm4hKGM4ftxfo
-pVR7KZy1pGvzD1jCUOvVg1kmNig2oG/c0cX64Y2e13zcTuKn/bFO4Ntey2cFga1J
-4G0qOBOkkwXFX75UC59QP/R3qWxsF+93V8rz1qjLoIIHj8X5Yjl2wEsbTR0C9l2B
-WwzNiHXIMrZb1tGwOe4Ndq6q9l9trPWj0iVhn6zsqGR0WtY3/x10RhW7bpVpfiuk
-+GvkDeS+xnh8cwcbKE+aMHbWJhCBvOkRXdaXlgP9/d1Iv3FcZg9YirfI8JSYf0Nw
-J85xzfZxds3cBGxkhJzfPHk2Gq4tofVRTGyrX4UWDaWrMpXTbhBUPc72psvlIKkC
-AwEAAaNvMG0wHQYDVR0OBBYEFF5RgDYc3WjotyWk9BbaJF3AVSiGMB8GA1UdIwQY
-MBaAFF5RgDYc3WjotyWk9BbaJF3AVSiGMA8GA1UdEwEB/wQFMAMBAf8wGgYDVR0R
-BBMwEYIJbG9jYWxob3N0hwR/AAABMA0GCSqGSIb3DQEBCwUAA4ICAQBwAPHm7hLD
-reUgvFgISTNSRomim+M9YNjFlBwdWEZOnInBPIK8n9pE8DzmSuTpSgQrRrl4l1QN
-9lGPMLTVIav9M/dHy+pQZ09tzuu7tq/zrN5lBwvPZCifUch1ML2IX0WTqtHM+bvD
-Eek0klhnXLz+xaFD6o7WsQrM+Ql53yHb0uQWoGvV9R2PoaMmHEyzBNsd2IbHXX8d
-+gfWxvA+PCV+d/s8PsOuz8sofLwtewY8pf+9xvXvnfPmjOdH2L6s1mpYx0SchnlN
-DAx4bsU8ye6MmlPF3yUZrc8CfL7wuAZchRQzR8NXhSxmO6e0jSMd/nRWLWCK4NKB
-ERBCcV8UBC2+EG/In7uRWDdq0bbeQAgvp78J4QiATIIcDn3txxkCIsNYfWX7fn0/
-HuYkrsTjW8/t6ePK/k2NWXZtXAihM7Bhi7LnGpnTzSEJn5GMb49BUlhQHSbZFyzO
-WVeGhYYesA7rnZata0WSyaAnJ9dPGGKOT/vsVMZvavEULxWXqrf4qd1W9HBJHmo4
-t4sSX23CmyKAyT4dTac1gzS8X1PJU2xzDvVC+had1Z4jE7107Wt12HaeYsh3vZng
-iqaDn3ysJIiIhuWU/CJBuYvBnwe/H4eOcS2ZPDuzPQxS8BYljx0saq3KKZ05fpeM
-NHpnSV0q8oK6nau5HzXJs0IUoWfVosoaww==
------END CERTIFICATE-----
